Devlop -> master

+++ src/migrations/1726757374843-AddDeletedAtToComment.ts
@@ -2,14 +2,20 @@ import { MigrationInterface, QueryRunner, TableColumn } from "typeorm";
 
 export class AddDeletedAtToComment1234567890123 implements MigrationInterface {
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.addColumn('comment', new TableColumn({
-            name: 'deleted_at',
-            type: 'timestamp',
-            isNullable: true,
-        }));
+        const hasColumn = await queryRunner.hasColumn('comment', 'deleted_at');
+        if (!hasColumn) {
+            await queryRunner.addColumn('comment', new TableColumn({
+                name: 'deleted_at',
+                type: 'timestamp',
+                isNullable: true,
+            }));
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.dropColumn('comment', 'deleted_at');
+        const hasColumn = await queryRunner.hasColumn('comment', 'deleted_at');
+        if (hasColumn) {
+            await queryRunner.dropColumn('comment', 'deleted_at');
+        }
     }
 }
